requirements.txt
streamlit
librosa
openai-whisper
spacy
pandas
numpy
soundfile
imageio-ffmpeg
langchain
langchain-openai
python-dotenv
plotly
torch
